# FunctionFrequency with finite state machine