### Current function:
```
BenchmarkFunctionFrequency-8    1784    686438 ns/op   221002 B/op    5021 allocs/op
PASS
ok  	github.io/somepackage/test/functionfrequency	1.296s
```

### With FSM:
```
BenchmarkFunctionFrequencyFSM-8    5376    213228 ns/op    7119 B/op    145 allocs/op
PASS
ok  	github.io/somepackage/test/functionfrequency	1.171s
```

#### Is your solution the most efficient way to solve a problem? Why?
This is a solution without using the go parser, just for functions lookup. Maybe not very perfect, but it works faster and more efficiency for memory.

#### Have you used any specific algorithm?
Finite state machine

#### What is time and space complexity of your solution?
O (N) except after search for sort functions